{{ cookiecutter.repo_name }}/bin/requirements.txt
# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
kedro
{% if cookiecutter.python_interpreter != 'python3' %}

# backwards compatibility
pathlib2
{% endif %}